refactor: formatValidationErrorメソッドをutilsに切り出し
