+++ src/utils/formatValidationError.ts
@@ -1,0 +1,22 @@
+const formatValidationError = (error: unknown) => {
+  if (
+    error &&
+    typeof error === "object" &&
+    "issues" in error &&
+    Array.isArray(error.issues)
+  ) {
+    const firstIssue = error.issues[0];
+    if (
+      firstIssue &&
+      typeof firstIssue === "object" &&
+      "message" in firstIssue &&
+      typeof firstIssue.message === "string"
+    ) {
+      return firstIssue.message;
+    }
+  }
+
+  return "入力内容を確認してください。";
+};
+
+export default formatValidationError;
